+++ src/EnumerablePrinter.Linq/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace EnumerablePrinter.Linq;
-
-public class Class1
-{
-
-}